Get Ensemble Samples

Get parameter values used in ensemble

get.ensemble.samples(
  ensemble.size,
  pft.samples,
  env.samples,
  method = "uniform",
  param.names = NULL,
  ...
)

Arguments

ensemble.size

number of runs in model ensemble

pft.samples

random samples from parameter distribution, e.g. from a MCMC chain

env.samples

env samples

method

the method used to generate the ensemble samples. Random generators: uniform, uniform with latin hypercube permutation. Quasi-random generators: halton, sobol, torus. Random generation draws random variates whereas quasi-random generation is deterministic but well equidistributed. Default is uniform. For small ensemble size with relatively large parameter number (e.g ensemble size < 5 and # of traits > 5) use methods other than halton.

param.names

a list of parameter names that were fitted either by MA or PDA, important argument, if NULL parameters will be resampled independently

...

Other arguments passed on to the sampling method

Value

matrix of (quasi-)random samples from trait distributions

Details

Returns a matrix of randomly or quasi-randomly sampled trait values to be assigned to traits over several model runs. given the number of model runs and a list of sample distributions for traits The model run is indexed first by model run, then by trait

Author

David LeBauer, Istem Fer