Sensitivity plot

Plot univariate response of model output to a trait parameter.

plot_sensitivity(
  sa.sample,
  sa.spline,
  trait,
  y.range = c(0, 50),
  median.i = 4,
  prior.sa.sample = NULL,
  prior.sa.spline = NULL,
  fontsize = list(title = 12, axis = 8),
  linesize = 1,
  dotsize = 2
)

Arguments

sa.sample

trait quantiles used in sensitivity analysis

sa.spline

spline function estimated from sensitivity analysis

trait

trait name for title

y.range

limits for y axis of plot

median.i

index of median value in sa.sample; median.i == which(as.numeric(rownames(sa.sample)) == 50)

prior.sa.sample

similar to sa.sample, but for prior distribution. If given, plots sensitivity for prior run

prior.sa.spline

similar to sa.spline, but for prior trait distribution.

fontsize

(optional) list with three arguments that can be set to vary the fontsize of the title, axis labels, and axis title in the sensitivity plots

linesize

passed to ggplot to set line thickness

dotsize

passed to ggplot to set point size

Value

object of class ggplot

Details

Plots for a single trait; called by plot_sensitivities to plot sensitivity plots for multiple traits.