Truncate spline at zero if...

Truncate spline at zero if P[x<0] < pnorm(-3) pnorm(-3) chosen as default value for min quantile because this is the default low end of range for the sensitivity analysis. This parameter could be determined based on minimum value in settings$sensitivity.analysis$quantiles

Usage

spline.truncate(x, min.quantile = stats::pnorm(-3))

Arguments

x

vector

min.quantile

threshold quantile for testing lower bound on variable

Value

either x or a vector with values < 0 converted to zero

Author

David LeBauer

Examples

set.seed(0)
x <- c(rgamma(998,1,1), rnorm(10))
min(x) # -0.5238
#> [1] -2.041882
min(PEcAn.uncertainty::spline.truncate(x))
#> [1] -2.041882