Write sensitivity analysis config files

Writes config files for use in sensitivity analysis.

Usage

write.sa.configs(
  defaults,
  quantile.samples,
  settings,
  model,
  clean = FALSE,
  write.to.db = TRUE
)

Arguments

defaults

named list with default parameter values

quantile.samples

list of lists supplied by get.sa.samples

settings

list of settings

model

name of model to be run

clean

logical: Delete any existing contents of the directory specified by settings$rundir before writing to it?

write.to.db

logical: Record this run to BETY? If TRUE, uses connection settings specified in settings$database

Value

list, containing $runs = data frame of runids, and $ensemble.id = the ensemble ID for these runs. Also writes sensitivity analysis configuration files as a side effect

Author

David LeBauer, Carl Davidson