Reads output of sensitivity analysis runs

Reads output of sensitivity analysis runs

Usage

read.sa.output(
  traits,
  quantiles,
  pecandir,
  outdir,
  pft.name = "",
  start.year,
  end.year,
  variable,
  sa.run.ids = NULL,
  per.pft = FALSE
)

Arguments

traits

model parameters included in the sensitivity analysis

quantiles

quantiles selected for sensitivity analysis

pecandir

specifies where pecan writes its configuration files

outdir

directory with model output to use in sensitivity analysis

pft.name

name of PFT used in sensitivity analysis (Optional)

start.year

first year to include in sensitivity analysis

end.year

last year to include in sensitivity analysis

variable

variables to be read from model output

sa.run.ids

list of run ids to read. If NULL, will look in `pecandir` for a file named `samples.Rdata` and read from that

per.pft

flag to determine whether we want SA on pft-specific variables

Value

dataframe with one col per quantile analysed and one row per trait, each cell is a list of AGB over time

Author

Ryan Kelly, David LeBauer, Rob Kooper, Mike Dietze, Istem Fer